python/requirements.txt
requests==2.21.0
oauthlib==3.0.1
requests_oauthlib==1.2.0
flask==1.0.2
Flask-WTF==0.14.2
PyJWT==1.7.1